% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owprm.R
\docType{data}
\name{ascorbic}
\alias{ascorbic}
\title{ascorbic data from Crowder and Hand}
\format{A data frame with 84 observations on the following 7 variables.}
\usage{
ascorbic
}
\description{
A dataset containing ascorbic acid measures on 12 patients
The data come from Table 3.3 in Crowder and Hand.
}
\keyword{datasets}

