% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owprm.R
\name{owprm.default}
\alias{owprm.default}
\title{Perform Occasions Within Phases repeated measures analysis.}
\usage{
\method{owprm}{default}(aovObj, ...)
}
\arguments{
\item{aovObj}{object returned from \code{aov}}

\item{...}{other arguments passed to function (not currently in use)

The \emph{fully saturated} \code{aov} object has the form \deqn{y_{ijk} = \pi_{i} + \Phi_{j} + o_{k(j)} + \pi \Phi_{ij} + \pi o_{ik(j)}}}
}
\value{
\code{owp_aovObj} A list containing an anova summary table
    of one line and a summary of the \code{aovObj}
}
\description{
Perform Occasions Within Phases repeated measures analysis.
}
\examples{
data(ascorbic)
aovObj <- aov(ascorbic.acid ~ patient*Phase*occ, data = ascorbic)
aovObj
## To avoid having warning messages sent to console
## suppressWarnings(owprm(aovObj))
owprm_aovObj <- owprm(aovObj)
owprm_aovObj
summary(owprm_aovObj)
## display p value in prettier form
options(digits = 4, scipen = 2)
owprm_aovObj
summary(owprm_aovObj)
}

