% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owprm.R
\name{owprm}
\alias{owprm}
\title{Occasions Within Phases Repeated Measures}
\usage{
owprm(aovObj, ...)
}
\arguments{
\item{aovObj}{object returned from \code{aov}}

\item{...}{other arguments passed to function (not currently in use)}
}
\description{
Occasions Within Phases Repeated Measures
}

