% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owprm.R
\name{print.summary.owprm}
\alias{print.summary.owprm}
\title{Print summary}
\usage{
\method{print}{summary.owprm}(x, ...)
}
\arguments{
\item{x}{\code{owprm} object}

\item{...}{other arguments to function}
}
\description{
Print summary
}

