% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owprm.R
\name{summary.owprm}
\alias{summary.owprm}
\title{Summary}
\usage{
\method{summary}{owprm}(object, ...)
}
\arguments{
\item{object}{\code{owprm} object}

\item{...}{other arguments to function}
}
\description{
Summary
}

